Clean up of source code, build, uml
Separated source code to different files, adjust build,
add class diagram for atrace
