+++ src/filesystem_impl.cpp
@@ -1,0 +1,87 @@
+/*
+ * Copyright (C) 2012 The Android Open Source Project
+ * Copyright (C) 2018 Roman Vaivod
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "filesystem_impl.h"
+
+#include <errno.h>  // errno
+#include <string.h> // strerror
+#include <unistd.h> // access
+#include <fcntl.h>  // creat, open
+
+void FileSystemImpl::set_errstream(FILE * errstream)
+{
+    this->errstream = errstream;
+}
+
+bool FileSystemImpl::fileExists(const char* filename)
+{
+    return access(filename, F_OK) != -1;
+}
+
+bool FileSystemImpl::fileIsWritable(const char* filename)
+{
+    return access(filename, W_OK) != -1;
+}
+
+bool FileSystemImpl::truncateFile(const char* path)
+{
+    // This uses creat rather than truncate because some of the debug kernel
+    // device nodes (e.g. k_ftraceFilterPath) currently aren't changed by
+    // calls to truncate, but they are cleared by calls to creat.
+    int traceFD = creat(path, 0);
+    if (traceFD == -1) {
+        fprintf(errstream, "error truncating %s: %s (%d)\n", path,
+            strerror(errno), errno);
+        return false;
+    }
+
+    close(traceFD);
+
+    return true;
+}
+
+bool FileSystemImpl::writeStr(const char* filename, const char* str)
+{
+    return _writeStr(filename, str, O_WRONLY);
+}
+
+bool FileSystemImpl::appendStr(const char* filename, const char* str)
+{
+    return _writeStr(filename, str, O_APPEND|O_WRONLY);
+}
+
+bool FileSystemImpl::_writeStr(const char* filename, const char* str, int flags)
+{
+    int fd = open(filename, flags);
+    if (fd == -1) {
+        fprintf(errstream, "error opening %s: %s (%d)\n", filename,
+                strerror(errno), errno);
+        return false;
+    }
+
+    bool ok = true;
+    ssize_t len = strlen(str);
+    if (write(fd, str, len) != len) {
+        fprintf(errstream, "error writing to %s: %s (%d)\n", filename,
+                strerror(errno), errno);
+        ok = false;
+    }
+
+    close(fd);
+
+    return ok;
+}
