+++ src/systemtime_impl.cpp
@@ -1,0 +1,31 @@
+/*
+ * Copyright (C) 2012 The Android Open Source Project
+ * Copyright (C) 2018 Roman Vaivod
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "systemtime_impl.h"
+
+#include "time.h"
+#include <utils/Timers.h>
+
+float SystemTimeImpl::get_monotonic() const
+{
+  return systemTime(CLOCK_MONOTONIC) / 1000000000.0f;
+}
+
+int64_t SystemTimeImpl::get_realtime() const
+{
+  return systemTime(CLOCK_REALTIME) / 1000000;
+}
