+++ src/tracingcategory.h
@@ -1,0 +1,42 @@
+/*
+ * Copyright (C) 2012 The Android Open Source Project
+ * Copyright (C) 2018 Roman Vaivod
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef LTTLWHWHT_TRACINGCATEGORY_H
+#define LTTLWHWHT_TRACINGCATEGORY_H
+
+#include "enablefile.h"
+
+#include <inttypes.h>
+#include <vector>
+
+class TracingCategory {
+  public:
+    // The name identifying the category.
+    const char* name;
+
+    // A longer description of the category.
+    const char* longname;
+
+    // The userland tracing tags that the category enables.
+    uint64_t tags;
+
+    std::vector<EnableFile> files;
+
+    bool is_enabled;
+};
+
+#endif // LTTLWHWHT_TRACINGCATEGORY_H
